moved lib.rs to stable.rs in wasm-bindgen-futures, updated during review
