+++ crates/futures/Cargo.toml
@@ -11,6 +11,7 @@ version = "0.3.25"
 edition = "2018"
 
 [dependencies]
+cfg-if = "0.1.9"
 futures = "0.1.20"
 js-sys = { path = "../js-sys", version = '0.3.25' }
 wasm-bindgen = { path = "../..", version = '0.2.48' }
@@ -20,7 +21,7 @@ lazy_static = { version = "1.3.0", optional = true }
 
 [target.'cfg(target_feature = "atomics")'.dependencies.web-sys]
 path = "../web-sys"
-version = "0.3.23"
+version = "0.3.24"
 features = [
     "MessageEvent",
     "Worker",
